api/pyproject.toml
[tool.pdm]
[tool.pdm.scripts]
server = "python3.11 -m uvicorn sampleboi_api.asgi:application --reload --host 0.0.0.0 --port 80"

# don't not treat this service as a package
[tool.setuptools]
py-modules = []

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Igor Bińkowski", email = "igorbinkowski1@gmail.com"},
]
dependencies = [
    "django>=4.2.1",
    "djangorestframework>=3.14.0",
    "uvicorn>=0.22.0",
    "psycopg2>=2.9.6",
    "psycopg2-binary>=2.9.6",
    "django-extensions>=3.2.1",
    "django-cors-headers>=3.14.0",
    "boto3>=1.26.129",
]
requires-python = ">=3.11"
license = {text = "MIT"}
